% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BigQuery_setup.R
\name{bigquery_setup_ui}
\alias{bigquery_setup_ui}
\title{BigQuery Setup UI}
\usage{
bigquery_setup_ui(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
The BigQuery Setup UI
}
\description{
This module is designed to guide a user through the process of authenticating with Google BigQuery. It is responsible for returning an authorization token, the user selected project,the user selected dataset, and a DBI connection to a BigQuery Dataset.
}
\keyword{internal}
