% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_chartreview.R
\name{chart_review}
\alias{chart_review}
\title{ReviewR Chart Review Tab}
\usage{
chart_review()
}
\value{
\link[shiny:renderUI]{shiny::renderUI} Output containing the Chart Review Tab
}
\description{
This function contains all of the elements that control the layout of
the Chart Review Tab.
}
\seealso{
Other layout: 
\code{\link{homepage}()},
\code{\link{patient_search}()},
\code{\link{setup}()}
}
\concept{layout}
\keyword{internal}
