% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_04_chart_review.R
\name{chart_review_tab}
\alias{chart_review_tab}
\title{ReviewR Chart Review Tab}
\usage{
chart_review_tab()
}
\value{
Chart Review Tab UI Output
}
\description{
This file contains all elements that are needed to render the Chart Review Tab
}
\keyword{internal}
