% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{concept}
\alias{concept}
\title{concept}
\format{
A data frame with 2355 rows and 10 variables:
\describe{
\item{concept_name}{\emph{character}}
\item{domain_id}{\emph{character}}
\item{vocabulary_id}{\emph{character}}
\item{concept_class_id}{\emph{character}}
\item{standard_concept}{\emph{character}}
\item{concept_code}{\emph{character}}
\item{valid_start_date}{\emph{Date}}
\item{valid_end_date}{\emph{Date}}
\item{invalid_reason}{\emph{character}}
\item{concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
concept
}
\description{
\subsection{Table Description}{

The Standardized Vocabularies contains records, or Concepts, that uniquely identify each fundamental unit of meaning used to express clinical information in all domain tables of the CDM. Concepts are derived from vocabularies, which represent clinical information across a domain (e.g. conditions, drugs, procedures) through the use of codes and associated descriptions. Some Concepts are designated Standard Concepts, meaning these Concepts can be used as normative expressions of a clinical entity within the OMOP Common Data Model and within standardized analytics. Each Standard Concept belongs to one domain, which defines the location where the Concept would be expected to occur within data tables of the CDM.

Concepts can represent broad categories (like ‘Cardiovascular disease’), detailed clinical elements (‘Myocardial infarction of the anterolateral wall’) or modifying characteristics and attributes that define Concepts at various levels of detail (severity of a disease, associated morphology, etc.).

Records in the Standardized Vocabularies tables are derived from national or international vocabularies such as SNOMED-CT, RxNorm, and LOINC, or custom Concepts defined to cover various aspects of observational data analysis.
}
}
\keyword{datasets}
