% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{concept_class}
\alias{concept_class}
\title{concept_class}
\format{
A data frame with 0 rows and 3 variables:
\describe{
\item{concept_class_id}{\emph{character}}
\item{concept_class_name}{\emph{character}}
\item{concept_class_concept_id}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_class }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
concept_class
}
\description{
\subsection{Table Description}{

The CONCEPT_CLASS table is a reference table, which includes a list of the classifications used to differentiate Concepts within a given Vocabulary. This reference table is populated with a single record for each Concept Class.
}
}
\keyword{datasets}
