% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{concept_synonym}
\alias{concept_synonym}
\title{concept_synonym}
\format{
A data frame with 5835 rows and 2 variables:
\describe{
\item{concept_synonym_name}{\emph{character}}
\item{concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_synonym }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
concept_synonym
}
\description{
\subsection{Table Description}{

The CONCEPT_SYNONYM table is used to store alternate names and descriptions for Concepts.
}
}
\keyword{datasets}
