% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\docType{data}
\name{db_module_template}
\alias{db_module_template}
\title{Database Module Template}
\format{
A character vector with 52 elements
}
\usage{
db_module_template
}
\description{
A character vector containing a database module template
}
\seealso{
Other Development Templates: 
\code{\link{db_function_all_patients_table_template}},
\code{\link{db_function_subject_table_template}}
}
\concept{Development Templates}
\keyword{datasets}
