% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\name{demo_sqlite_setup_server}
\alias{demo_sqlite_setup_server}
\title{Demo SQLite Setup Server}
\usage{
demo_sqlite_setup_server(id)
}
\arguments{
\item{id}{The Module namespace}
}
\value{
Demo SQLite connection variables
}
\description{
Demo SQLite Setup Server
}
\keyword{internal}
