% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\name{demo_sqlite_setup_ui}
\alias{demo_sqlite_setup_ui}
\title{Demo SQLite Setup UI}
\usage{
demo_sqlite_setup_ui(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
The Demo SQLite Setup UI
}
\description{
This module is designed to guide a user through the process of authenticating your database
}
