% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_add_database_module}
\alias{dev_add_database_module}
\title{Develop A Database Module}
\usage{
dev_add_database_module(mod_name = NULL, display_name = NULL)
}
\arguments{
\item{mod_name}{\emph{Required}. A string, denoting the module suffix eg: 'mariadb'}

\item{display_name}{\emph{Required}. A string, denoting the module display name eg: 'MariaDB Server'.
This is the 'user viewable' name that will appear in the database module selector dropdown.}
}
\value{
A .R file populated with a database module skeleton
}
\description{
This function will create a database module skeleton  with
required elements already populated, based on user inputs.
Common database module packages are imported automatically,
but developers should add imports to the roxygen skeleton as
necessary to both the UI and server functions to collect
user info and create a DBI connection object, respectively.
}
