% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_add_google_client_id}
\alias{dev_add_google_client_id}
\title{Add Google Client ID}
\usage{
dev_add_google_client_id(file_path)
}
\arguments{
\item{file_path}{Path to a "Desktop" Google Client ID JSON on your local system.}
}
\description{
This function will move a Google desktop Client ID json from Google Cloud into the appropriate
directory on your computer to be used by ReviewR in place of the built in credentials.
}
\seealso{
Other Development Functions: 
\code{\link{dev_add_database_module}()},
\code{\link{dev_add_datamodel}()}
}
\concept{Development Functions}
