% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{drug_exposure}
\alias{drug_exposure}
\title{drug_exposure}
\format{
A data frame with 363 rows and 22 variables:
\describe{
\item{drug_exposure_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{drug_concept_id}{\emph{integer}}
\item{drug_exposure_start_date}{\emph{Date}}
\item{drug_exposure_start_datetime}{\emph{POSIXct}}
\item{drug_exposure_end_date}{\emph{Date}}
\item{drug_exposure_end_datetime}{\emph{POSIXct}}
\item{verbatim_end_date}{\emph{Date}}
\item{drug_type_concept_id}{\emph{integer}}
\item{stop_reason}{\emph{character}}
\item{refills}{\emph{integer}}
\item{quantity}{\emph{numeric}}
\item{days_supply}{\emph{integer}}
\item{sig}{\emph{character}}
\item{route_concept_id}{\emph{integer}}
\item{lot_number}{\emph{character}}
\item{provider_id}{\emph{integer}}
\item{visit_occurrence_id}{\emph{integer}}
\item{drug_source_value}{\emph{character}}
\item{drug_source_concept_id}{\emph{integer}}
\item{route_source_value}{\emph{character}}
\item{dose_unit_source_value}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_exposure }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
drug_exposure
}
\description{
\subsection{Table Description}{

The Device domain captures information about a person’s exposure to a foreign physical object or instrument which is used for diagnostic or therapeutic purposes through a mechanism beyond chemical action. Devices include implantable objects (e.g. pacemakers, stents, artificial joints), medical equipment and supplies (e.g. bandages, crutches, syringes), other instruments used in medical procedures (e.g. sutures, defibrillators) and material used in clinical care (e.g. adhesives, body material, dental material, surgical material).
}

\subsection{User Guide}{

The distinction between Devices or supplies and Procedures are sometimes blurry, but the former are physical objects while the latter are actions, often to apply a Device or supply.
}

\subsection{ETL Conventions}{

Source codes and source text fields mapped to Standard Concepts of the Device Domain have to be recorded here.
}
}
\keyword{internal}
