% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{drug_strength}
\alias{drug_strength}
\title{drug_strength}
\format{
A data frame with 475035 rows and 9 variables:
\describe{
\item{amount_value}{\emph{numeric}}
\item{numerator_value}{\emph{numeric}}
\item{denominator_value}{\emph{numeric}}
\item{box_size}{\emph{integer}}
\item{valid_start_date}{\emph{Date}}
\item{valid_end_date}{\emph{Date}}
\item{invalid_reason}{\emph{character}}
\item{ingredient_concept_id}{\emph{integer}}
\item{drug_concept_id}{\emph{integer}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_strength }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
drug_strength
}
\description{
\subsection{Table Description}{

The DRUG_STRENGTH table contains structured content about the amount or concentration and associated units of a specific ingredient contained within a particular drug product. This table is supplemental information to support standardized analysis of drug utilization.
}
}
\keyword{internal}
