% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dt_2_viewer}
\alias{dt_2_viewer}
\title{DT to Viewer}
\usage{
dt_2_viewer(.data, file = NULL)
}
\arguments{
\item{.data}{A tibble/data frame containing the desired data to save}

\item{file}{File location with extension}
}
\value{
HTML file
}
\description{
Save a DT::datatable as a self contained html file to display as a prompt mid function
}
\keyword{internal}
