% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_get_concepts.R
\name{get_all_concept}
\alias{get_all_concept}
\title{OMOP Get Concepts}
\usage{
get_all_concept(
  table_map,
  db_connection,
  concept_table,
  concept_id,
  concept_name,
  table,
  joinable_id,
  table_concept_id,
  col_name
)
}
\arguments{
\item{table_map}{tibble containing standard CDM mapped to user tables}

\item{db_connection}{DBI connection object}

\item{concept_table}{Standard CDM concept table name}

\item{concept_id}{Standard CDM concept id field}

\item{concept_name}{Standard CDM concept name field}

\item{table}{Table that needs concepts retrieved}

\item{joinable_id}{What variable is joinable between the concept table and the desired table}

\item{table_concept_id}{The table concept id}

\item{col_name}{Would you like to rename the retrieved concept?}
}
\value{
The desired OMOP concepts based on the user data model for all subjects
}
\description{
OMOP Get Concepts
}
\examples{
\dontrun{
gender_concepts <- get_all_concept(table_map = table_map,  db_connection = db_connection,  
    concept_table = 'concept', concept_id = 'concept_id',  concept_name = 'concept_name',  
    table = 'person', joinable_id = 'person_id', table_concept_id = 'gender_concept_id',  
    col_name = 'Gender')
provider_concepts <- get_all_concept(table_map = table_map,  db_connection = db_connection,  
    concept_table = 'provider', concept_id = 'provider_id',  concept_name = 'provider_name',  
    table = 'person', joinable_id = 'person_id', table_concept_id = 'provider_id',  
    col_name = 'Provider') 
}
}
\keyword{internal}
