% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_omop_get_concepts.R
\name{get_subject_concept}
\alias{get_subject_concept}
\title{OMOP Get Subject Concepts}
\usage{
get_subject_concept(
  table_map,
  db_connection,
  concept_table,
  concept_id,
  concept_name,
  table,
  joinable_id,
  table_concept_id,
  col_name,
  subject_id_field,
  selected_subject
)
}
\arguments{
\item{table_map}{tibble containing standard CDM mapped to user tables}

\item{db_connection}{DBI connection object}

\item{concept_table}{Standard CDM concept table name}

\item{concept_id}{Standard CDM concept id field}

\item{concept_name}{Standard CDM concept name field}

\item{table}{Table that needs concepts retrieved}

\item{joinable_id}{What variable is joinable between the concept table and the desired table}

\item{table_concept_id}{The table concept id}

\item{col_name}{Would you like to rename the retrieved concept?}

\item{subject_id_field}{wWhich field contains the subject id?}

\item{selected_subject}{Which subject would you like to retrieve concepts for}
}
\value{
The desired OMOP concepts based on the user data model for a particular subject
}
\description{
OMOP Get Subject Concepts
}
\keyword{internal}
