% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_homepage.R
\name{homepage}
\alias{homepage}
\title{ReviewR Homepage Tab}
\usage{
homepage()
}
\value{
\link[shiny:renderUI]{shiny::renderUI} Output containing the Homepage Tab
}
\description{
This function contains all of the elements that control the layout of
the Homepage Tab.
}
\seealso{
Other layout: 
\code{\link{chart_review}()},
\code{\link{patient_search}()},
\code{\link{setup}()}
}
\concept{layout}
\keyword{internal}
