% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BigQuery_setup.R
\name{installed_app}
\alias{installed_app}
\title{Installed App}
\usage{
installed_app()
}
\value{
An Invisible OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}}
}
\description{
Invisibly returns an OAuth app.
}
\keyword{internal}
