% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_tables_mimic.R
\name{mimic_table_all_patients}
\alias{mimic_table_all_patients}
\alias{mimic_table_admissions}
\alias{mimic_table_callout}
\alias{mimic_table_chart_events}
\alias{mimic_table_cpt_events}
\alias{mimic_table_diagnoses_icd}
\alias{mimic_table_drg_codes}
\alias{mimic_table_icu_stays}
\alias{mimic_table_lab_events}
\alias{mimic_table_microbiology_events}
\alias{mimic_table_note_events}
\alias{mimic_table_prescriptions}
\alias{mimic_table_procedure_events}
\alias{mimic_table_procedures_icd}
\alias{mimic_table_services}
\alias{mimic_table_transfers}
\title{MIMIC Tables}
\usage{
mimic_table_all_patients(table_map, db_connection)

mimic_table_admissions(table_map, db_connection, subject_id)

mimic_table_callout(table_map, db_connection, subject_id)

mimic_table_chart_events(table_map, db_connection, subject_id)

mimic_table_cpt_events(table_map, db_connection, subject_id)

mimic_table_diagnoses_icd(table_map, db_connection, subject_id)

mimic_table_drg_codes(table_map, db_connection, subject_id)

mimic_table_icu_stays(table_map, db_connection, subject_id)

mimic_table_lab_events(table_map, db_connection, subject_id)

mimic_table_microbiology_events(table_map, db_connection, subject_id)

mimic_table_note_events(table_map, db_connection, subject_id)

mimic_table_prescriptions(table_map, db_connection, subject_id)

mimic_table_procedure_events(table_map, db_connection, subject_id)

mimic_table_procedures_icd(table_map, db_connection, subject_id)

mimic_table_services(table_map, db_connection, subject_id)

mimic_table_transfers(table_map, db_connection, subject_id)
}
\arguments{
\item{table_map}{tibble containing a the cdm that most closely matches the user's database and a map of standard tables to user tables}

\item{db_connection}{Connection info received from the database setup module}

\item{subject_id}{The selected subject}
}
\description{
Collection of functions to create pre-arranged views of MIMIC patient data for ReviewR.
}
\keyword{internal}
