% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abstraction_setup.R
\name{abstraction_setup_ui}
\alias{abstraction_setup_ui}
\alias{mod_abstraction_setup_server}
\title{Module Selector}
\usage{
abstraction_setup_ui(id)

mod_abstraction_setup_server(id, subject_id)
}
\arguments{
\item{id}{}
}
\description{
A Shiny module to select from available ReviewR abstraction modules
}
\keyword{internal}
