% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chart_abstraction_setup_module.R
\name{chart_abstraction_setup_ui}
\alias{chart_abstraction_setup_ui}
\alias{chart_abstraction_setup_logic}
\title{Chart Abstraction Setup Module}
\usage{
chart_abstraction_setup_ui(id)

chart_abstraction_setup_logic(input, output, session, abstraction_selection)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{abstraction_selection}{Which abstraction backend has been selected?}
}
\description{
Chart Abstraction Setup Module
}
\keyword{internal}
