% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_bigquery_module.R
\name{bq_auth_ui}
\alias{bq_auth_ui}
\alias{bq_project_auth_logic}
\alias{bq_dataset_auth_logic}
\alias{bq_initialize}
\title{BigQuery Module}
\usage{
bq_auth_ui(id)

bq_project_auth_logic(input, output, session)

bq_dataset_auth_logic(input, output, session, bq_project)

bq_initialize(input, output, session, bq_project, bq_dataset, disconnect)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{bq_project}{A BigQuery Project ID}

\item{bq_dataset}{A BigQuery Dataset ID}

\item{disconnect}{An action button press}
}
\description{
This module is designed to guide a user through the process of authenticating with Google BigQuery. It is responsible for returning an authorization token, the user selected project,the user selected dataset, and a DBI connection to a BigQuery Dataset.
}
\keyword{internal}
