% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chart_abstraction_save_module.R
\name{instrument_complete_ui}
\alias{instrument_complete_ui}
\alias{instrument_complete_logic}
\title{Chart Abstraction Save Module}
\usage{
instrument_complete_ui(id)

instrument_complete_logic(
  input,
  output,
  session,
  rc_instrument,
  instrumentData,
  previousData,
  all_instruments,
  instrument_selection,
  subjectID
)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{rc_instrument}{A reactive tibble containing the chart abstraction instrument}

\item{instrumentData}{A reactive tibble containing the instrument schema}

\item{previousData}{A reactive tibble containing any previously entered data for the currently selected subject}

\item{all_instruments}{All instruments in the current abstraction project}

\item{instrument_selection}{The currently selected abstraction instrument}

\item{subjectID}{The currently selected subject identifier}
}
\description{
This module is responsible for rendering the "Save Abstraction" dialogue. It ensures that all required question types are answered before presenting the user with a dropdown to 'complete' the review. A save button is also provided.
}
\keyword{internal}
