% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_setup.R
\name{database_setup_ui}
\alias{database_setup_ui}
\alias{mod_database_setup_server}
\title{Module Selector}
\usage{
database_setup_ui(id)

mod_database_setup_server(id)
}
\arguments{
\item{id}{}
}
\description{
A Shiny module to select from available ReviewR database modules
}
\keyword{internal}
