% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_module.R
\name{db_setup_ui}
\alias{db_setup_ui}
\alias{db_select_logic}
\alias{db_connect_logic}
\title{Database Setup Module}
\usage{
db_setup_ui(id)

db_select_logic(input, output, session)

db_connect_logic(input, output, session, db_type, db_disconnect)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{db_type}{Database type selection}

\item{db_disconnect}{An Action button press}
}
\description{
A Shiny module to step a user through the process of selecting a database and calling the appropriate module based on selection.
}
\keyword{internal}
