% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datamodel_setup.R
\name{datamodel_detection_ui}
\alias{datamodel_detection_ui}
\alias{mod_datamodel_detection_server}
\title{Data Model Detection Module}
\usage{
datamodel_detection_ui(id)

mod_datamodel_detection_server(
  id,
  database_vars,
  navigation_vars,
  parent_session
)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{navigation_vars}{Navigation variables returned from mod_navigation}

\item{parent_session}{the parent environment of this module}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{db_connection}{Connection info received from the database setup module}
}
\description{
This module is designed to connect to a user supplied database, compare it with
known common data models and determine the most likely version of the user's database.
}
\keyword{internal}
