% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigation.R
\name{navigation_message}
\alias{navigation_message}
\alias{mod_navigation_server}
\title{Patient Navigation}
\usage{
navigation_message(id)

mod_navigation_server(
  id,
  database_vars,
  datamodel_vars,
  abstract_vars,
  parent_session
)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{database_vars}{Database variables returned from user selected database setup module}

\item{datamodel_vars}{Datamodel variables returned from mod_datamodel_setup}

\item{abstract_vars}{Abstraction variables returned from user selected abstraction module}

\item{parent_session}{the parent environment of this module}
}
\description{
This module will render the datatable on the 'Patient Search' tab containing all patients in the cohort.
The selected patient in the DT is kept in sync with the 'Chart Review' tab.
}
\keyword{internal}
