% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_offline_abstraction_module.R
\name{offline_setup_ui}
\alias{offline_setup_ui}
\alias{offline_setup_select_logic}
\alias{offline_setup_logic}
\alias{offline_abs_connected_ui}
\alias{offline_connected_logic}
\alias{offline_abs_config_ui}
\title{Offline Abstraction Module}
\usage{
offline_setup_ui(id)

offline_setup_select_logic(input, output, session)

offline_setup_logic(input, output, session, selection)

offline_abs_connected_ui(id)

offline_connected_logic(input, output, session, abstraction_vars)

offline_abs_config_ui(id)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{selection}{Existing or new session?}

\item{abstraction_vars}{Abstraction setup variables}
}
\description{
This module contains all of the Offline setup and instrumnet upload/download components.
}
\keyword{internal}
