% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_patient_chart_module.R
\name{omop_chart_review_ui}
\alias{omop_chart_review_ui}
\alias{omop_chart_review_logic}
\alias{mimic_chart_review_ui}
\alias{mimic_chart_review_logic}
\title{Patient Chart Module}
\usage{
omop_chart_review_ui(id)

omop_chart_review_logic(
  input,
  output,
  session,
  table_map,
  db_connection,
  subject_id
)

mimic_chart_review_ui(id)

mimic_chart_review_logic(
  input,
  output,
  session,
  table_map,
  db_connection,
  subject_id
)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{table_map}{tibble containing a the cdm that most closely matches the user's database and a map of standard tables to user tables}

\item{db_connection}{Connection info received from the database setup module}

\item{subject_id}{The selected subject}
}
\description{
This module is designed to guide a user through the process of authenticating with Google BigQuery. It is responsible for returning an authorization token, the user selected project,the user selected dataset, and a DBI connection to a BigQuery Dataset.
}
\keyword{internal}
