% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_redcap_module.R
\name{redcap_connect_ui}
\alias{redcap_connect_ui}
\alias{redcap_connect_logic}
\alias{redcap_initialize_logic}
\alias{redcap_connected_ui}
\alias{redcap_connected_logic}
\alias{redcap_instrument_config_ui}
\alias{redcap_instrument_select_logic}
\alias{redcap_instrument_config_logic}
\alias{redcap_instrument_config_reviewer_logic}
\alias{rc_instrument_configured_ui}
\alias{rc_instrument_configured_logic}
\alias{redcap_instrument_ui}
\alias{redcap_instrument_logic}
\alias{upload_redcap_logic}
\title{REDCap Module}
\usage{
redcap_connect_ui(id)

redcap_connect_logic(input, output, session)

redcap_initialize_logic(input, output, session, rc_url, rc_token)

redcap_connected_ui(id)

redcap_connected_logic(input, output, session, connect, rc_project_info)

redcap_instrument_config_ui(id)

redcap_instrument_select_logic(
  input,
  output,
  session,
  rc_connect_press,
  rc_connection
)

redcap_instrument_config_logic(
  input,
  output,
  session,
  rc_connection,
  instruments,
  instrument_selection
)

redcap_instrument_config_reviewer_logic(
  input,
  output,
  session,
  rc_instrument,
  rc_identifier,
  rc_connection
)

rc_instrument_configured_ui(id)

rc_instrument_configured_logic(
  input,
  output,
  session,
  rc_instrument_info,
  selected_instrument
)

redcap_instrument_ui(id)

redcap_instrument_logic(
  input,
  output,
  session,
  rc_connection,
  instruments,
  instrument_selection,
  rc_instrument,
  rc_identifier,
  rc_reviewer,
  rc_selected_reviewer,
  subject_id,
  reviewr_upload_btn,
  reviewr_connect_btn
)

upload_redcap_logic(
  input,
  output,
  session,
  rc_connection,
  rc_recordID,
  rc_instrument,
  instrumentData,
  previousData,
  currentSubject,
  rc_upload_btn_press,
  abstraction_complete,
  abstraction_complete_val,
  all_instruments,
  instrument_selection
)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{rc_url}{A character string containing the REDCap API URL for your institution ex: https://redcap.ucdenver.edu/api/}

\item{rc_token}{A password string representing your REDCap API token}

\item{connect}{A press of the REDCap connect button from the Setup Tab}

\item{rc_project_info}{The project information assosciated with your REDCap API key}

\item{rc_connect_press}{A press of the REDCap connect button from the Setup Tab}

\item{rc_connection}{A REDCap API connection object}

\item{instruments}{All REDCap instruments assosciated with the project API key}

\item{instrument_selection}{The user selected REDCap Instrument}

\item{rc_instrument}{The user selected REDCap Instrument}

\item{rc_identifier}{The field that contains the REDCap Identifier}

\item{rc_instrument_info}{Information about the configured REDCap Instrument}

\item{selected_instrument}{The currently selected REDCap Instrument}

\item{rc_reviewer}{The REDCap field that holds reviewer names}

\item{rc_selected_reviewer}{The selected reviewer from the list of previous reviewers (if present)}

\item{subject_id}{The currently selected patient identifier}

\item{reviewr_upload_btn}{A button press of the REDCap upload button}

\item{reviewr_connect_btn}{A button press of the REDCap Connect button}

\item{rc_recordID}{The REDCap Record ID for the currently selected subject}

\item{instrumentData}{The user entered data currently present in the REDCap instrument}

\item{previousData}{Previous Abstraction "Instrument Complete" Value}

\item{currentSubject}{Abstraction data assosciated with the currently selected subject}

\item{rc_upload_btn_press}{A press of the Upload to REDCap button}

\item{abstraction_complete}{Person readable abstraction complete value}

\item{abstraction_complete_val}{Machine readable abstraction complete value}

\item{all_instruments}{All data abstraction instruments in the project}
}
\description{
This module will assist with selecting a patient. It keeps the patient search tab data table up todate, server side renders the select input on the patient chart tab, and displays in context information about the selected patient.
}
\keyword{internal}
