% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_sqlite_module.R
\name{sqlite_ui}
\alias{sqlite_ui}
\alias{sqlite_logic}
\title{SQLite Module}
\usage{
sqlite_ui(id)

sqlite_logic(input, output, session, disconnect)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{disconnect}{A press of the database disconnect button}
}
\value{
A UI for connecting to the demo data


}
\description{
This module will create a SQLite database in memory to allow a user to preview the functionality of ReviewR.
}
\keyword{internal}
