% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{observation}
\alias{observation}
\title{observation}
\format{
A data frame with 184 rows and 17 variables:
\describe{
\item{observation_id}{\emph{integer}}
\item{person_id}{\emph{integer}}
\item{observation_concept_id}{\emph{integer}}
\item{observation_date}{\emph{Date}}
\item{observation_datetime}{\emph{POSIXct}}
\item{observation_type_concept_id}{\emph{integer}}
\item{value_as_number}{\emph{numeric}}
\item{value_as_string}{\emph{character}}
\item{value_as_concept_id}{\emph{integer}}
\item{qualifier_concept_id}{\emph{integer}}
\item{unit_concept_id}{\emph{integer}}
\item{provider_id}{\emph{integer}}
\item{visit_occurrence_id}{\emph{integer}}
\item{observation_source_value}{\emph{character}}
\item{observation_source_concept_id}{\emph{integer}}
\item{unit_source_value}{\emph{character}}
\item{qualifier_source_value}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#observation }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
observation
}
\description{
\subsection{Table Description}{

The OBSERVATION table captures clinical facts about a Person obtained in the context of examination, questioning or a procedure. Any data that cannot be represented by any other domains, such as social and lifestyle facts, medical history, family history, etc. are recorded here.
}

\subsection{User Guide}{

Observations differ from Measurements in that they do not require a standardized test or some other activity to generate clinical fact. Typical observations are medical history, family history, the stated need for certain treatment, social circumstances, lifestyle choices, healthcare utilization patterns, etc. If the generation clinical facts requires a standardized testing such as lab testing or imaging and leads to a standardized result, the data item is recorded in the MEASUREMENT table. If the clinical fact observed determines a sign, symptom, diagnosis of a disease or other medical condition, it is recorded in the CONDITION_OCCURRENCE table. Valid Observation Concepts are not enforced to be from any domain though they still should be Standard Concepts.
}

\subsection{ETL Conventions}{

Records whose Source Values map to any domain besides Condition, Procedure, Drug, Measurement or Device should be stored in the Observation table. Observations can be stored as attribute value pairs, with the attribute as the Observation Concept and the value representing the clinical fact. This fact can be a Concept (stored in VALUE_AS_CONCEPT), a numerical value (VALUE_AS_NUMBER), a verbatim string (VALUE_AS_STRING), or a datetime (VALUE_AS_DATETIME). Even though Observations do not have an explicit result, the clinical fact can be stated separately from the type of Observation in the VALUE_AS_* fields. It is recommended for Observations that are suggestive statements of positive assertion should have a value of ‘Yes’ (concept_id=4188539), recorded, even though the null value is the equivalent.
}
}
\keyword{internal}
