% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_patientsearch.R
\name{patient_search}
\alias{patient_search}
\title{ReviewR Patient Search Tab}
\usage{
patient_search()
}
\value{
\link[shiny:renderUI]{shiny::renderUI} Output containing the Patient Search Tab
}
\description{
This function contains all of the elements that control the layout of
the Patient Search Tab.
}
\seealso{
Other layout: 
\code{\link{chart_review}()},
\code{\link{homepage}()},
\code{\link{setup}()}
}
\concept{layout}
\keyword{internal}
