% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PostgreSQL_setup.R
\name{postgresql_setup_server}
\alias{postgresql_setup_server}
\title{mod_PostgreSQL_setup Server Function}
\usage{
postgresql_setup_server(id)
}
\arguments{
\item{id}{The PostgreSQL Setup UI}
}
\value{
PostgreSQL connection variables and DBI connection object.
}
\description{
mod_PostgreSQL_setup Server Function
}
\keyword{internal}
