% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PostgreSQL_setup.R
\name{postgresql_setup_ui}
\alias{postgresql_setup_ui}
\title{PostgreSQL Setup UI}
\usage{
postgresql_setup_ui(id)
}
\arguments{
\item{id}{The Module namespace}
}
\value{
The PostgreSQL Setup UI
}
\description{
This module is designed to guide a user through the process of authenticating with a PostgreSQL database. It is responsible for creating a DBI connection object to a PostgreSQL database.
}
\keyword{internal}
