% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\name{redcap_instrument_ui}
\alias{redcap_instrument_ui}
\title{REDCap Instrument UI}
\usage{
redcap_instrument_ui(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
A Shiny Representation of REDCap Instrument
}
\description{
REDCap Instrument UI
}
\keyword{internal}
