% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\name{redcap_server}
\alias{redcap_server}
\title{REDCap Server}
\usage{
redcap_server(id, subject_id)
}
\arguments{
\item{id}{The module namespace}

\item{subject_id}{A reactive expression containing a subject identifier}
}
\value{
REDCap connection variables and project information
}
\description{
REDCap Server
}
\keyword{internal}
