% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\name{redcap_setup_ui}
\alias{redcap_setup_ui}
\title{REDCap Setup UI}
\usage{
redcap_setup_ui(id)
}
\arguments{
\item{id}{The module namespace}
}
\value{
The REDCap Setup UI
}
\description{
REDCap Setup UI
}
