% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\docType{data}
\name{redcap_survey_complete}
\alias{redcap_survey_complete}
\title{REDCap Survey Complete}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{redcap_survey_complete_names}{The human readable "Survey Complete" Responses}
\item{redcap_survey_complete_values}{REDCap API values for "Survey Complete" Responses}
...
}
}
\usage{
redcap_survey_complete
}
\description{
A dataset containing valid REDCap "Survey Complete" Values.
}
\keyword{datasets}
