% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_redcap_survey_complete_values.R
\docType{data}
\name{redcap_survey_complete_values}
\alias{redcap_survey_complete_values}
\title{REDCap Survey Complete Values}
\format{
A named list containing 3 items:
\describe{
\item{Incomplete}{REDCap Survey Incomplete}
\item{Unverified}{REDCap Survey needs verification}
\item{Complete}{REDCap Survey answered completely}
...
}
}
\usage{
redcap_survey_complete_values
}
\description{
A named list containing REDCap "Survey Complete" Values.
}
\keyword{datasets}
