% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_redcap_widget_map.R
\docType{data}
\name{redcap_widget_map}
\alias{redcap_widget_map}
\title{REDCap Widget Map}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{redcap_field_type}{A REDCap Question Type}
\item{redcap_field_val}{Custom REDCap Question Type Validation}
\item{reviewr_redcap_widget_function}{ReviewR function to use when mapping to native Shiny widget}
...
}
}
\usage{
redcap_widget_map
}
\description{
A dataset that maps REDCap question types and common validations
to native shiny widgets through custom functions.
}
\keyword{datasets}
