% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReviewR-package.R
\docType{package}
\name{ReviewR-package}
\alias{ReviewR}
\alias{ReviewR-package}
\title{ReviewR: A Light-Weight, Portable Tool for Reviewing Individual Patient Records}
\description{
A portable Shiny tool to explore patient-level electronic health record data and perform chart review in a single integrated framework. This tool supports browsing clinical data in many different formats including multiple versions of the 'OMOP' common data model as well as the 'MIMIC-III' data model. In addition, chart review information is captured and stored securely via the Shiny interface in a 'REDCap' (Research Electronic Data Capture) project using the 'REDCap' API. See the 'ReviewR' website for additional information, documentation, and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reviewr.thewileylab.org/}
  \item \url{https://github.com/thewileylab/ReviewR/}
  \item Report bugs at \url{https://github.com/thewileylab/ReviewR/issues}
}

}
\author{
\strong{Maintainer}: David Mayer \email{david.mayer@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6056-9771}{ORCID})

Authors:
\itemize{
  \item Laura Wiley \email{laura.wiley@cuanschutz.edu} (\href{https://orcid.org/0000-0001-6681-9754}{ORCID})
  \item Luke Rasmussen \email{luke.rasmussen@northwestern.edu} (\href{https://orcid.org/0000-0002-4497-8049}{ORCID})
}

Other contributors:
\itemize{
  \item The Wiley Lab [copyright holder, funder]
}

}
\keyword{internal}
