% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{package}
\name{ReviewR}
\alias{ReviewR}
\title{ReviewR: A light-weight, portable tool for reviewing individual patient records}
\description{
ReviewR is a portable Shiny tool to help you explore data across different
data models. Within ReviewR, you can browse patient data stored in either
the OMOP or MIMIC-III data model. In addition to viewing patient data, you
may also connect to a REDCap project to perform a chart review.
}
\section{Development}{
NA
}

