% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_REDCap.R
\name{safe_exportRecords}
\alias{safe_exportRecords}
\title{Safe Export Records}
\usage{
safe_exportRecords(rc_con, rc_field_names)
}
\arguments{
\item{rc_con}{A REDCap API Connection Object}

\item{rc_field_names}{The field names for a REDCap instrument}
}
\value{
A data frame containing existing REDCap records, or an empty data frame with the structure of what the records would look like
}
\description{
Export REDCap Records in a REDCap Project. Sometimes, records don't exist (empty instrument). Use field names to create empty data structure.
}
\keyword{internal}
