% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BigQuery_setup.R
\name{safe_file.exists}
\alias{safe_file.exists}
\title{Safe File Exists}
\usage{
safe_file.exists(...)
}
\arguments{
\item{...}{character vectors, containing file names or paths}
}
\value{
Logical, true/false if file path is provided or NULL if not supplied
with any input.
}
\description{
A "safe" wrapper around \link[base:files]{base::file.exists} that returns a FALSE if no file
path is supplied as an argument, instead of an error.
}
\keyword{internal}
