% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_setup.R
\name{setup}
\alias{setup}
\title{ReviewR Setup Tab}
\usage{
setup()
}
\value{
\link[shiny:renderUI]{shiny::renderUI} Output containing the Setup Tab
}
\description{
This function contains all of the elements that control the layout of
the Setup Tab.
}
\seealso{
Other layout: 
\code{\link{chart_review}()},
\code{\link{homepage}()},
\code{\link{patient_search}()}
}
\concept{layout}
\keyword{internal}
