% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_02_setup.R
\name{setup_tab}
\alias{setup_tab}
\title{ReviewR Setup Tab}
\usage{
setup_tab()
}
\value{
Setup Tab UI Output
}
\description{
This file contains the UI elements that are needed to render the Setup Tab and configure ReviewR. Each element consists of a module, which is rendered within it's own box element on the setup tab.
}
\examples{
output$setup_tab <- setup_tab()
}
