% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{synPUF}
\alias{synPUF}
\title{synPUF}
\format{
A data frame with 23 rows and 2 variables:
\describe{
\item{table_name}{\emph{character}: The table name in OMOP v5.2.2}
\item{table_data}{\emph{list}: The table data, in OMOP v5.2.2}
...
}
}
\source{
\url{https://github.com/thewileylab/synPUF/releases/tag/0.0.1.10}

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF.html}

\url{https://github.com/OHDSI/CommonDataModel/blob/v5.2.2/PostgreSQL/OMOP\%20CDM\%20ddl\%20-\%20PostgreSQL.sql}

\url{https://www.mtsamples.com/}
}
\usage{
synPUF
}
\description{
\subsection{Overview}{

This dataset contains a 10 person subset of the CMS 2008-2010 Data Entrepreneurs’
Synthetic Public Use File (DE-SynPUF) from OHDSI.
}

\subsection{Details}{
\itemize{
\item \href{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF.html}{Information on the SynPUF data source is available here}
\item \href{https://github.com/OHDSI/CommonDataModel/blob/v5.2.2/PostgreSQL/OMOP\%20CDM\%20ddl\%20-\%20PostgreSQL.sql}{CDM 5.2.2 DDL for the OHDSI supported DBMSs is available here}
\item \href{https://www.mtsamples.com/}{Notes Obtained from MTSamples.com}
}
}
}
\keyword{internal}
