% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_sqlite.R
\docType{data}
\name{vocabulary}
\alias{vocabulary}
\title{vocabulary}
\format{
A data frame with 0 rows and 5 variables:
\describe{
\item{vocabulary_id}{\emph{character}}
\item{vocabulary_name}{\emph{character}}
\item{vocabulary_reference}{\emph{character}}
\item{vocabulary_version}{\emph{character}}
\item{vocabulary_concept_id}{\emph{character}}
...
}
}
\source{
\url{https://ohdsi.github.io/CommonDataModel/cdm531.html#vocabulary }

\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
vocabulary
}
\description{
\subsection{Table Description}{

The VOCABULARY table includes a list of the Vocabularies collected from various sources or created de novo by the OMOP community. This reference table is populated with a single record for each Vocabulary source and includes a descriptive name and other associated attributes for the Vocabulary.
}
}
\keyword{internal}
