% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_01_welcome_tab.R
\name{welcome_tab}
\alias{welcome_tab}
\title{ReviewR Welcome Tab}
\usage{
welcome_tab()
}
\value{
Welcome Tab UI Output
}
\description{
This file contains all the UI elements that are needed to render the Welcome Tab, which contains basic information about using ReviewR.
}
\keyword{internal}
