% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_write_sqlite_table.R
\name{write_sqlite_table}
\alias{write_sqlite_table}
\title{Write SQLite Table}
\usage{
write_sqlite_table(table_location, table_name, db_connection)
}
\arguments{
\item{table_location}{Where the rda file resides on disk}

\item{table_name}{the desired of the table contained on disk}

\item{db_connection}{connection information for SQLite DB}
}
\description{
A function to write an RDA file to a SQLite database.
}
\keyword{internal}
