% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\docType{data}
\name{db_function_all_patients_table_template}
\alias{db_function_all_patients_table_template}
\title{Database Table Function: All Patients Table Template}
\format{
A character vector with 22 elements
}
\usage{
db_function_all_patients_table_template
}
\description{
A character vector containing a function template for creating the
'All Patients' table as displayed on the "Patient Search" Tab
}
\seealso{
Other Development Templates: 
\code{\link{db_function_subject_table_template}},
\code{\link{db_module_template}}
}
\concept{Development Templates}
\keyword{datasets}
