% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\docType{data}
\name{db_function_subject_table_template}
\alias{db_function_subject_table_template}
\title{ReviewR Database Function: Subject Table Template}
\format{
A character vector with 13 elements
}
\usage{
db_function_subject_table_template
}
\description{
A character vector containing a function template for creating the
'Subject Filtered' tables as displayed on the "Chart Review" Tab
}
\keyword{datasets}
