% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\docType{data}
\name{db_module_template}
\alias{db_module_template}
\title{ReviewR Database Module Template}
\format{
A character vector with 51 elements
}
\usage{
db_module_template
}
\description{
A character vector containing a database module template
}
\keyword{datasets}
