% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_add_data_model}
\alias{dev_add_data_model}
\title{Develop Data Model Table Functions}
\usage{
dev_add_data_model(csv)
}
\arguments{
\item{csv}{\emph{Required}. The file path of a CSV file containing a data model schema}
}
\value{
A .R file populated with basic database table functions
}
\description{
This function will assist in adding a new supported data model to
ReviewR. A schema file, stored as a CSV will be added to the
namespace, such that the database can be identified by the data model
detection module. Additionally, a database_tables.R file will
be created and opened in R/ with basic table skeletons created
based on the schema stored in the user supplied CSV.
}
\seealso{
Other Development Functions: 
\code{\link{dev_add_database_module}()}
}
\concept{Development Functions}
