% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_add_datamodel}
\alias{dev_add_datamodel}
\title{Develop Datamodel Functions}
\usage{
dev_add_datamodel(csv, all_patients_table, patient_identifier_field)
}
\arguments{
\item{csv}{The file path of a CSV file containing a datamodel CDM}

\item{all_patients_table}{In this CDM, which table contains a listing of all patients?}

\item{patient_identifier_field}{In this CDM, which field denotes the patient identifier?}
}
\description{
This function will assist in adding a new supported datamodel to
ReviewR. A CDM file, stored as a CSV will be added to the
namespace, such that the CDM can be identified by the datamodel
detection module. Additionally, a database_tables.R file will
be created and opened in R/ with basic table skeletons created
based on the CDM stored in the user supplied CSV.
}
