% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_add_google_client_id}
\alias{dev_add_google_client_id}
\title{Add Google Client ID}
\usage{
dev_add_google_client_id(file_path)
}
\arguments{
\item{file_path}{Path to a "Desktop" Google Client ID JSON on your local system.}
}
\description{
Add Google Client ID
}
