% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dev_database_module}
\alias{dev_database_module}
\title{Develop Database Module}
\usage{
dev_database_module(mod_name = NULL, display_name = NULL)
}
\arguments{
\item{mod_name}{A string, denoting the module suffix eg: 'mariadb'}

\item{display_name}{A string, denoting the module display name eg: 'MariaDB'}
}
\value{

}
\description{
This function will assist in providing a database module skeleton for ReviewR
}
