% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dt_2_viewer}
\alias{dt_2_viewer}
\title{DT to Viewer}
\usage{
dt_2_viewer(.data, file = NULL)
}
\arguments{
\item{.data}{A tibble/data frame containing the desired data to save}

\item{file}{\emph{Optional}. Manually define file path (with .html extension) for HTML representation of DT}
}
\value{
Temporary HTML file to be displayed in the RStudio Viewer Pane
}
\description{
Save a DT::datatable as a self contained HTML file to display for display in the RStudio Viewer Pane
}
\keyword{internal}
