% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewr_dev.R
\name{dt_2_viewer}
\alias{dt_2_viewer}
\title{DT to Viewer}
\usage{
dt_2_viewer(.data, file = NULL)
}
\arguments{
\item{.data}{A \link[dplyr:reexports]{dplyr::tibble} containing the desired data to save}

\item{file}{\emph{Optional}. Manually define file path (with .html extension)
for HTML representation of DT}
}
\value{
This function returns a temporary HTML file displayed in the
RStudio Viewer Pane
}
\description{
Save a temporary \link[DT:datatable]{DT::datatable} as a self contained HTML file to display
in the RStudio Viewer Pane. Used to provided users with choices
when prompted for action by a dev function.
}
\keyword{internal}
