% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_homepage.R
\name{homepage}
\alias{homepage}
\title{ReviewR Homepage Tab}
\usage{
homepage()
}
\value{
Homepage Tab UI Output
}
\description{
This file contains all the UI elements that are needed to render the Homepage Tab, which contains basic information about using ReviewR.
}
\keyword{internal}
