% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datamodel_setup.R
\name{data_model_detection_ui}
\alias{data_model_detection_ui}
\alias{mod_data_model_detection_server}
\title{Data Model Detection Module}
\usage{
data_model_detection_ui(id)

mod_data_model_detection_server(id, database_vars)
}
\arguments{
\item{id}{The namespace id for the UI output}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{db_connection}{Connection info received from the database setup module}
}
\description{
This module is designed to connect to a user supplied database, compare it with known common data models and determine the most likely version of the user's database.
}
\keyword{internal}
