% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_tables.R
\name{omop_table_all_patients}
\alias{omop_table_all_patients}
\alias{omop_table_condition_era}
\alias{omop_table_condition_occurrence}
\alias{omop_table_death}
\alias{omop_table_device_exposure}
\alias{omop_table_dose_era}
\alias{omop_table_drug_era}
\alias{omop_table_drug_exposure}
\alias{omop_table_measurement}
\alias{omop_table_note}
\alias{omop_table_observation}
\alias{omop_table_observation_period}
\alias{omop_table_payer_plan_period}
\alias{omop_table_procedure_occurrence}
\alias{omop_table_specimen}
\alias{omop_table_visit_occurrence}
\title{OMOP Tables}
\usage{
omop_table_all_patients(table_map, db_connection)

omop_table_condition_era(table_map, db_connection, subject_id)

omop_table_condition_occurrence(table_map, db_connection, subject_id)

omop_table_death(table_map, db_connection, subject_id)

omop_table_device_exposure(table_map, db_connection, subject_id)

omop_table_dose_era(table_map, db_connection, subject_id)

omop_table_drug_era(table_map, db_connection, subject_id)

omop_table_drug_exposure(table_map, db_connection, subject_id)

omop_table_measurement(table_map, db_connection, subject_id)

omop_table_note(table_map, db_connection, subject_id)

omop_table_observation(table_map, db_connection, subject_id)

omop_table_observation_period(table_map, db_connection, subject_id)

omop_table_payer_plan_period(table_map, db_connection, subject_id)

omop_table_procedure_occurrence(table_map, db_connection, subject_id)

omop_table_specimen(table_map, db_connection, subject_id)

omop_table_visit_occurrence(table_map, db_connection, subject_id)
}
\arguments{
\item{table_map}{tibble containing a the cdm that most closely matches the user's database and a map of standard tables to user tables}

\item{db_connection}{Connection info received from the database setup module}

\item{subject_id}{The selected subject}
}
\description{
Collection of functions to create pre-arranged views of OMOP patient data for ReviewR.
}
\keyword{internal}
