% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_patient_search.R
\name{patient_search}
\alias{patient_search}
\title{ReviewR Patient Search Tab}
\usage{
patient_search()
}
\value{
Patient Search Tab UI Output
}
\description{
This file contains all elements that are needed to render the Patient Search Tab
}
\keyword{internal}
