% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_connection.R
\name{redcap_connection}
\alias{redcap_connection}
\title{REDCap Connection}
\usage{
redcap_connection(url, token)
}
\arguments{
\item{url}{The API URL for your institution's REDCap instance}

\item{token}{The API token for your REDCap project}
}
\value{
If the URL and token are correct, return a redcapAPI connection object. Else, return diagnostic error.
}
\description{
A 'safe' wrapper for redcapAPI::redcapConnection(). Will return diagnostic error codes in case incorrect URL or token are provided.
}
\keyword{internal}
