% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_redcap_instrument.R
\name{reviewr_textInput}
\alias{reviewr_textInput}
\alias{reviewr_dateInput}
\alias{reviewr_dropdown}
\alias{reviewr_truefalse}
\alias{reviewr_yesno}
\alias{reviewr_radio}
\alias{reviewr_checkbox}
\alias{reviewr_notes}
\alias{reviewr_integer}
\alias{render_redcap}
\title{Render REDCap Instrument}
\usage{
reviewr_textInput(id, field_label, value = NULL, placeholder = NULL, ...)

reviewr_dateInput(id, field_label, value = NULL, ...)

reviewr_dropdown(id, field_label, required, choices, value = NULL, ...)

reviewr_truefalse(id, field_label, required, value = NULL, ...)

reviewr_yesno(id, field_label, required, value = NULL, ...)

reviewr_radio(id, field_label, required, choices, value = NULL, ...)

reviewr_checkbox(id, field_label, choices, value = NULL, ...)

reviewr_notes(id, field_label, value = NULL, ...)

reviewr_integer(id, field_label, value = NULL, ...)

render_redcap(
  reviewr_type,
  field_name,
  field_label,
  required,
  choices,
  current_subject_data = NULL
)
}
\arguments{
\item{id}{Unique REDCap question identifier}

\item{field_label}{Question text, with formatting}

\item{value}{Default value or previous data if question has previously been answered}

\item{placeholder}{Placeholder text to help a reviewer decide how to answer the question}

\item{...}{Any additional parameters}

\item{required}{Is this a required REDCap question type?}

\item{choices}{REDCap choices for the question.}

\item{current_subject_data}{Previous REDCap data on the current subject}
}
\description{
Collection of functions to map REDCap question types to native Shiny widgets.
}
\keyword{internal}
