% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reviewr_datatable.R
\name{reviewr_datatable}
\alias{reviewr_datatable}
\title{Render ReviewR Datatable}
\usage{
reviewr_datatable(.data)
}
\arguments{
\item{.data}{A local tibble or dataframe to be rendered in the ReviewR UI}
}
\value{
return a DT with custom options
}
\description{
Render ReviewR Datatable
}
\keyword{internal}
