% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigation.R
\name{reviewr_datatable}
\alias{reviewr_datatable}
\title{ReviewR Datatable}
\usage{
reviewr_datatable(.data, dom = "ftip", column_filter = "top", search_term = "")
}
\arguments{
\item{.data}{A local tibble or data frame to be rendered in the ReviewR UI}

\item{dom}{Define the table control elements to appear on the page and in what order}

\item{column_filter}{Where to display individual column filters. Valid entries are: 'top','bottom','none'}

\item{search_term}{A string or regular expression used as a filter for patient data}
}
\value{
return a DT with custom options
}
\description{
ReviewR Datatable
}
\keyword{internal}
