% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ReviewR Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{A list of options to pass to golem}
}
\value{
No return value, called to start the ReviewR Shiny Application!
}
\description{
Run the ReviewR Application
}
