% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_setup.R
\name{setup}
\alias{setup}
\title{ReviewR Setup Tab}
\usage{
setup()
}
\value{
Setup Tab UI Output
}
\description{
This file contains the UI elements that are needed to render the Setup Tab and configure ReviewR. Each element consists of a module, which is rendered within it's own box element on the setup tab.
}
\keyword{internal}
