% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_supported_datamodels.R
\docType{data}
\name{supported_datamodels}
\alias{supported_datamodels}
\title{ReviewR Supported Database Schemas}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{file_path}{Where schema was imported from}
\item{datamodel}{Data model name, OMOP or MIMIC}
\item{model_version}{Version of the data model}
\item{data}{Nested database schemas, including included table and field mappings}
...
}
}
\source{
\url{https://github.com/OHDSI/CommonDataModel/}

\url{https://github.com/MIT-LCP/mimic-code}
}
\usage{
supported_datamodels
}
\description{
A dataset containing data model information along with the corresponding
version and schema.
}
\keyword{datasets}
