% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datamodel_setup.R
\name{user_field}
\alias{user_field}
\title{user_field}
\usage{
user_field(table_map, desired_cdm_table, desired_cdm_field)
}
\arguments{
\item{table_map}{A reactive tibble, generated by mod_datamodel_setup containing user database tables and fields mapped to the determined CDM}

\item{desired_cdm_table}{In the desired CDM, which table do you want?}

\item{desired_cdm_field}{In the desired CDM, which field do you want?}
}
\value{
character object containing the user database field pertaining to the omop field you want, from the omop table specified
}
\description{
user_field
}
\examples{
\dontrun{
user_field(table_map = table_map, desired_cdm_table = 'person', desired_cdm_field = 'person_id')
}
}
\keyword{internal}
