% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_datamodel_setup.R
\name{user_table}
\alias{user_table}
\title{user_table}
\usage{
user_table(table_map, db_con, desired_cdm_table)
}
\arguments{
\item{table_map}{A reactive tibble, generated by mod_datamodel_setup containing user database tables and fields mapped to the determined CDM}

\item{db_con}{A reactive DBI connection object that is created through user interaction with the Setup Tab}

\item{desired_cdm_table}{In the desired CDM, which table do you want?}
}
\value{
tbl(db_con, user_table) the user table that corresponds to the standard data model table you want
}
\description{
user_table
}
\examples{
\dontrun{
user_table(table_map = table_map, db_con = db_connection, desired_cdm_table = 'person')
}
}
\keyword{internal}
