% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_helper.R
\name{user_table}
\alias{user_table}
\title{user_table}
\usage{
user_table(table_map, connection_info, desired_cdm_table)
}
\arguments{
\item{table_map}{A reactive tibble, generated by mod_datamodel_setup containing user database tables and fields mapped to the determined CDM}

\item{connection_info}{A reactive DBI connection object that is created through user interaction with the Setup Tab}

\item{desired_cdm_table}{In the desired CDM, which table do you want?}
}
\value{
tbl(connection_info, user_table) the user table that corresponds to the omop table you want
}
\description{
user_table
}
\examples{
\dontrun{
user_table(table_map = table_map, connection_info = db_connection, desired_cdm_table = 'person')
}
}
\keyword{internal}
